requirements.txt
alembic==1.13.1
anyio==4.3.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.5
passlib==1.7.4
playwright==1.42.0
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
pyee==11.0.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
sqlmodel==0.0.16
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.28.0
watchfiles==0.21.0
websockets==12.0
wkhtmltopdf==0.2
wrapt==1.16.0
